requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
chromadb>=0.4.18
pyyaml>=6.0.1
tqdm>=4.66.1
cachetools>=5.3.2
numpy>=1.24.0
